feat: restructure engagement creation based on who is being engaged
