+++ app/Http/Controllers/EngagementController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers;
 use App\Enums\EngagementFormat;
 use App\Enums\EngagementRecruitment;
 use App\Enums\ProvinceOrTerritory;
+use App\Http\Requests\StoreEngagementFormatRequest;
 use App\Http\Requests\StoreEngagementLanguagesRequest;
-use App\Http\Requests\StoreEngagementOutreachRequest;
 use App\Http\Requests\StoreEngagementRecruitmentRequest;
 use App\Http\Requests\StoreEngagementRequest;
 use App\Http\Requests\UpdateEngagementLanguagesRequest;
@@ -50,7 +50,6 @@ class EngagementController extends Controller
     {
         return view('engagements.create', [
             'project' => $project,
-            'formats' => Options::forEnum(EngagementFormat::class)->toArray(),
         ]);
     }
 
@@ -70,18 +69,24 @@ class EngagementController extends Controller
 
         flash(__('Your engagement has been created.'), 'success');
 
-        return redirect(localized_route('engagements.show-outreach-selection', $engagement));
+        $redirect = match ($engagement->who) {
+            'organization' => localized_route('engagements.manage', $engagement),
+            default => localized_route('engagements.show-format-selection', $engagement),
+        };
+
+        return redirect($redirect);
     }
 
-    public function showOutreachSelection(Engagement $engagement): View
+    public function showFormatSelection(Engagement $engagement): View
     {
-        return view('engagements.show-outreach-selection', [
+        return view('engagements.show-format-selection', [
             'project' => $engagement->project,
             'engagement' => $engagement,
+            'formats' => Options::forEnum(EngagementFormat::class)->toArray(),
         ]);
     }
 
-    public function storeOutreach(StoreEngagementOutreachRequest $request, Engagement $engagement): RedirectResponse
+    public function storeFormat(StoreEngagementFormatRequest $request, Engagement $engagement): RedirectResponse
     {
         $engagement->fill($request->validated());
         $engagement->save();
@@ -90,12 +95,7 @@ class EngagementController extends Controller
 
         $engagement = $engagement->fresh();
 
-        $redirect = match ($engagement->who) {
-            'organization' => localized_route('engagements.manage', $engagement),
-            default => localized_route('engagements.show-recruitment-selection', $engagement),
-        };
-
-        return redirect($redirect);
+        return redirect(localized_route('engagements.show-recruitment-selection', $engagement));
     }
 
     public function showRecruitmentSelection(Engagement $engagement): View
