--- app/Http/Requests/StoreEngagementOutreachRequest.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class StoreEngagementOutreachRequest extends FormRequest
-{
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    public function rules(): array
-    {
-        return [
-            'who' => 'required|in:individuals,organization',
-        ];
-    }
-}
