+++ tests/RequestFactories/StoreEngagementRequestFactory.php
@@ -3,6 +3,7 @@
 namespace Tests\RequestFactories;
 
 use App\Models\Project;
+use function Pest\Faker\faker;
 use Worksome\RequestFactories\RequestFactory;
 
 class StoreEngagementRequestFactory extends RequestFactory
@@ -11,8 +12,8 @@ class StoreEngagementRequestFactory extends RequestFactory
     {
         return [
             'project_id' => Project::factory(),
-            'name' => ['en' => 'Workshop'],
-            'format' => 'workshop',
+            'name' => ['en' => 'Workshop '.faker()->randomNumber(5)],
+            'who' => 'individuals',
             'ideal_participants' => 25,
             'minimum_participants' => 15,
             'paid' => true,
